Better support between PS versions (WIP)
Streamline the surcharge fee handling to not use products as there is just too many pitfalls with that approach. PS does not support adding additional fees to an order, but changing the order total amounts and displaying information about the added surcharge on the order detail pages, and invoice should be good enough.