Fix wrong total for free shipping discount
When there is a free shipping discount voucher, the total is not calculated correctly.

Shipping is added to seller_costs, but the shipping discount is subtracted from order_total. If by chance the cost of the products is smaller than the shipping cost, then order_total becomes negative, and generates an error. The fix subtracts the shipping discount amount from the seller_costs, and add the shipping discount in the list as type shipping.